requirements.txt
glfw~=2.1.0
pyopengl~=3.1.5
